(MTWTV-1591) snapshot boot on single core . Root cause: Becasue api emmc_read offset parameter is signed long, but the original is unsinged long. And when emmc_read call our mmc_bread, the parameter offset is unsigned long long.When the bit32 is 1, mistake will come. Solution: Because the emmc size is 4G, so we change the emmc_read paramter to unsigned long.
---- mtk usage begin ---
MTK_Submitter: jjian.zhou
MTK_MP_CL: 1918583
MTK_RLS_CL: 1918589
MTK_CR_ID: DTV00634425
---- mtk usage end ---
.

+++ uboot/u-boot-2011.12/common/cmd_lg/cmd_misc.c
@@ -1124,7 +1124,7 @@ int storage_init(void)
 int storage_read(uint32_t offset, size_t len, void* buf)
 {
 #ifdef CC_EMMC_BOOT
-	return emmc_read((off_t)offset, (size_t)len, (u_char *)buf);
+	return emmc_read((uint32_t)offset, (size_t)len, (u_char *)buf);
 #else
 	return nand_read(&nand_info[0], (off_t)offset, (size_t *)&len, (u_char *)buf);
 #endif
